Исправления перед бета-тестом

+++ src/main.py
@@ -2,13 +2,13 @@ import sys, os, subprocess, psutil
 
 from generator import *
 from executor import *
-from parse_json import analyze
-from graphics import *
 
 from PyQt5.QtWidgets import *
 from mainWindow import *
 from optimizationWindow import *
 from resultsWindow import *
+from graphics import *
+from parse_json import jsonAnalyzer
 
 
 curDir = f'/home/{os.getlogin()}/PEAS/src'
@@ -30,6 +30,7 @@ class mainWindow(QMainWindow, Ui_MainWindow):
     
     # функция выбора и добавления директории до библиотеки
     def get_library_dir(self):
+
         library_dir = QFileDialog.getOpenFileName(self)[0]
         self.textEdit_1.append("Directory saved.")
 
@@ -61,72 +62,52 @@ class mainWindow(QMainWindow, Ui_MainWindow):
         h_file.close()
 
     # функция вызова окна оптимизации
-    def optimization(self):
-        window2 = OptimizationWindow()
-        window2.exec_()
+    #def optimization(self):
+        #window2 = OptimizationWindow()
+        #window2.exec_()
 
     def clear_tmp(self, need_to_clear):
         for tmp in need_to_clear:
             try: os.remove(f'{parDir}/{tmp}.cpp')
             except FileNotFoundError:
-                print('No cpp generated')
-        try: 
-            for tmpSub in os.listdir(tmpDir):
-                #print(f'{tmpDir}/{tmpSub}')
-                for fileName in os.listdir(f'{tmpDir}/{tmpSub}'):
-                        os.remove(f'{tmpDir}/{tmpSub}/{fileName}')
-                os.rmdir(f'{tmpDir}/{tmpSub}')
-        except FileNotFoundError:
-            print('No binaries in tmp')
-        
-        #try:
-        #    os.remove(f'{parDir}/hPath.txt')
-        #    os.remove(f'{parDir}/LibraryPath.txt')
-        #except FileNotFoundError:
-        #    print('No PATH files presented')
-
-        
-            
+                print('generated .cpp removed')
+            try: 
+                for tmpSub in os.listdir(tmpDir):
+                    #print(f'{tmpDir}/{tmpSub}')
+                    for fileName in os.listdir(f'{tmpDir}/{tmpSub}'):
+                        if not fileName.endswith('json'):
+                            #print('need to remove', fileName)
+                            os.remove(f'{tmpDir}/{tmpSub}/{fileName}')
+            except FileNotFoundError:
+                print('generated binaries removed')
 
     # функция сборки и вывода окна с результатами
     def results(self):
         # чтение пути до h файлов для generator и builder
-        
         h_file = open(f'{parDir}/hPath.txt', "r")
         h_common_way = h_file.read()
         h_way = h_common_way[0:h_common_way.rfind("/") + 1]
         h_file.close()
         # python3 generator.py /home/vadim/PEAS/src/
         mem_req = generator(parDir, h_way, h_common_way[h_common_way.rfind("/")+1: len(h_common_way)-1])
- 
-        
 
         # Чтение пути до библиотеки для builder
         library_file = open(f'{parDir}/LibraryPath.txt', "r")
         lib_way = library_file.read()
         library_file.close()
-        #
-        subprocess.run(['python3', f'{curDir}/builder_threads.py', lib_way, h_way, nanobench])
-        #
-        execute(mem_req)
-
-        scrapped_data = analyze()
-        #
-
-        need_to_clear = [tmp[0] for tmp in mem_req]
-        self.clear_tmp(need_to_clear)
+        #python3 builder.py /home/vadim/PEAS/src/lsm.py /home/vadim/PEAS/src/
+        if not (h_file and lib_way) :
+            subprocess.run(['python3', f'{curDir}/builder_threads.py', lib_way, h_way, nanobench])
+            execute(mem_req)
         
+        need_to_clear = [tmp[0] for tmp in mem_req]
+        need_to_clear.sort()
 
-
-        # Вывод окна с результатами
-        file_name = 'testsquare'
-        test_names = [file_name+'O0', file_name+'O1', file_name+'O2', file_name+'O3', file_name+'Ofast']
-        test_value1 = [100, 200, 300, 400, 500]
-        test_value2 = [100, 200, 300, 400, 500]
-        test_value3 = [100, 200, 300, 400, 500]
-        testsquare1 = Graph(test_names, test_value1, test_value2, test_value3)
-        testsquare1.create_graph()    
-
+        window3 = ResultsWindow()
+        window3.initUI(need_to_clear)
+        window3.exec_()
+        self.clear_tmp(need_to_clear)
+'''
 class OptimizationWindow(QDialog, Ui_Dialog1):
     def __init__(self):
         super().__init__()
@@ -134,17 +115,48 @@ class OptimizationWindow(QDialog, Ui_Dialog1):
         self.initUI()
 
     def initUI(self):
-        
         pass
+'''
 
 class ResultsWindow(QDialog, Ui_Dialog2):
     def __init__(self):
         super().__init__()
         self.setupUi(self)
-        self.initUI()
 
-    def initUI(self):
-        pass
+    def initUI(self, item_list):
+        for i in item_list:
+            self.listWidget.addItem(i)
+        self.okButton.clicked.connect(self.build_graph)
+
+    # Функция построения графиков
+    def build_graph(self):
+        data = jsonAnalyzer()
+        # Словарь данных
+        data_dict = data.get_res()
+
+        # Список всех названий тестов
+        all_test_names = list(data_dict.keys())
+        all_test_names.sort()
+
+        # Списки значений для построения графиков
+        all_test_ops = []
+        all_test_ipc = []
+        all_test_time = []
+
+        for i in all_test_names:
+            all_test_ops.append(round(data_dict[i]['op/s'], 2)) # 'op/s'
+            all_test_ipc.append(round(data_dict[i]['IPC'], 2))
+            all_test_time.append(data_dict[i]['median(elapsed)'])
+        
+        list_graph = []
+        for i in range(int(len(all_test_names)/5)):
+            list_graph.append(Graph(all_test_names[5*i:5*i+5], all_test_ops[5*i:5*i+5],all_test_ipc[5*i:5*i+5], all_test_time[5*i:5*i+5]))
+        
+        # Построение графиков
+        list_graph[self.listWidget.currentRow()].create_graph() # Убрать + 1, если не совпадают
+
+        self.hide()
+
 
 if __name__ == '__main__':
     app = QtWidgets.QApplication(sys.argv)
